Refactor pages into its own directory
